requirements.txt
streamlit==1.25.0
numpy==1.21.2
Pillow==8.4.0
tensorflow==2.6.0
opencv-python